plugins/mewma/pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "drift-mewma"
version = "2.0.0"
description = "MEWMA (Multivariate EWMA) drift detection plugin for drift-framework"
requires-python = ">=3.11"
dependencies = [
    "drift-framework>=0.1,<1.0",
    "scipy>=1.10",
]

[project.entry-points."drift_framework.plugins"]
mewma = "drift_mewma:register"

[tool.setuptools.packages.find]
include = ["drift_mewma*"]

[tool.setuptools.package-data]
drift_mewma = [
    "web/templates/**/*",
    "web/static/**/*",
]
